Add zip + deploy step to deploy_app

+++ app/deploy_app.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 set -x -e
 
+if [ $# -ne 1 ]; then
+    echo "Usage: deploy_app.sh (debug|release)"
+    exit 1
+fi
+
 cd ../build-$1
 make clean
+make qmake
 rm -rf antimony.app
 
 make
@@ -28,3 +34,11 @@ install_name_tool -change /usr/local/Frameworks/Python.framework/Versions/3.4/Py
 
 cd ../../..
 cp -r fab antimony.app/Contents/Frameworks/Python.framework/Versions/3.4/lib/python3.4/fab
+
+cp ../README.md .
+tar -cvzf antimony.tar.gz antimony.app README.md
+rm README.md
+
+if [ `whoami` = "mkeeter" ]; then
+    scp antimony.tar.gz mattkeeter.com:mattkeeter.com/projects/antimony/antimony.tar.gz
+fi
